+++ pkg/goanalysis/pkgerrors/errors.go
@@ -1,4 +1,4 @@
-package packages
+package pkgerrors
 
 import (
 	"errors"
