+++ pkg/goanalysis/pkgerrors/util.go
@@ -1,4 +1,4 @@
-package packages
+package pkgerrors
 
 import (
 	"fmt"
