+++ pkg/goanalysis/pkgerrors/util_test.go
@@ -1,4 +1,4 @@
-package packages
+package pkgerrors
 
 import (
 	"testing"
